start component refactor, added camera component
